use try_into for a fallible conversion to hashing::Digest (#9994)
### Problem

There is a conversion using the `From` trait from the protobuf Digest to a `Result<hashing::Digest, String>`. This conversion is failable so should use TryFrom instead.

Motivation:  I was modifying the `fs` crate to switch away from futures 0.1 and this change seemed to help some of the type inference, but mainly improved my understanding of the code since it was more clear where a `Result` type was appearing.

I've pulled it out here for ease of review.

### Solution

Use try_into for these conversions.

### Result

Tests pass.