Wait for ongoing runs to complete before shutting down pantsd (#10052)
### Problem

As described in #8200, the errors when pantsd shuts down during a run are not particularly helpful.

### Solution

When pantsd has chosen to exit (due to either of the `PantsService` instances exiting: generally the `SchedulerService` which watches the pythonpath and other invalidation globs for changes), wait for any ongoing connections to complete before exiting.

### Result

The server will wait for ongoing runs to complete before exiting. Fixes #8200.