disable append-only cache in remoting (#10077)
### Problem

The append-only cache code currently adds a platform property to RE requests to specify that an append-only cache should be mounted when executing the request (if supported).

We discovered that this does not interact well with REv2 servers that require that workers match on **all** platform properties (e.g. Buildbarn) or that require values (and not just keys) be known ahead of time even if they can match on a subset of properties (e.g. Buildfarm). Adding a platform property requires server configuration to either add additional workers (Buildbarn) or specify all possible values (Buildfarm, for example container-image).

### Solution

Disable the append-only caching in remoting by not setting the platform property.

### Result

Should prevent the need to hack around on the server-side for now to allow Pants to remote with Buildbarn.