Control workunit levels at `@rule` execution time (#10053)
### Problem

We want `@rule`s to be able to potentially modify the level of a workunit they are associated with at rule execution time.

### Solution

This commit introduces a new Python class called `EngineAware`, which defines an abstract method `level()` returning an Optional LogLevel. If the type returned by a rule subclasses `EngineAware`, the engine will use the output of that method to potentially override the level set on the workunit associated with running the rule. 

For now this framework is only used to allow rules to override their own workunit level, but we expect that in the future this class will be used to implement more pieces of rule metadata.