Fix the `coverage` field not working with v2 (#10082)
Previously, the `coverage` field would not work if the specified module lived in a different folder than the test. 

The issue is that we weren't using the `FileTracer` mechanism, which is necessary for Coverage to know which files Pants "owns". We were using `find_executable_files`, but we need `FileTracer` too. (We still need `find_executable_files` so that files which are not encountered during test execution still count as "owned" by this plugin.)

This PR also makes two small improvements:
* Don't merge data if there's only one test result. This avoids unnecessary work.
* If `-ldebug`, turn on some Coverage debug output. This is how v1 behaved, and it seems useful to preserve.

[ci skip-rust-tests]
[ci skip-jvm-tests]