Add report type to FilesystemCoverageReport
Work towards: https://github.com/pantsbuild/pants/issues/9968

Currently, looking at FilesystemCoverageReport the type of report (xml, html and in the future raw) is not specified.
Attaching that information is useful if we want to have rules and other mechansims (work units) that handle rule artifacts to run logic against this data type.