remoting: use debug level for remoting sub-workunits (#10098)
### Problem

The workunits generated by remoting for ensure_action_uploaded and run_execute_request were created with the default info level, which means they get "completed" statuses printed to the console. This is too verbose.

Example:

```
17:57:56.11 [INFO] Completed: ensure_action_uploaded
```

### Solution

Switch those workunits to debug level. This PR also adds a helper to WorkunitMetadata to make it easier to construct a default instance with an alternate level.

Fixes a bug in the final metadata where it always set error.

### Result

Less verbose.