Add support for MyPy to Pants v2 (#10132)
This is not a perfect implementation yet. See https://github.com/pantsbuild/pants/issues/10131 for remaining TODOs. But, this gives an initial implementation to build off of.

A key decision of this PR is to add MyPy to the `lint` goal, rather than a new `typecheck` goal. Users shared feedback that they prefer this, as it's neat to run all your linters in parallel. Technically, MyPy is a linter, only a supercharged one.

To activate, add `pants.backend.python.lint.mypy` to `backend_packages2`.

[ci skip-rust-tests]
[ci skip-jvm-tests]