Set dynamic-ui default based on CI (#10140)
### Problem

Several different continuous integration systems (including Travis and CircleCI) expose a tty to programs that execute within them, but doesn't correctly handle the dynamic UI screen updating, resulting in garbage being printed to the console in CI contexts.

### Solution

Pants now sets the default value of the `--dynamic-ui` option to False if there is an environment variable called `CI` with a value of `true`, which several popular CI systems do (including Travis and CircleCI), so the dynamic UI will be turned off by default in CI without requiring manual disabling. The dynamic UI can still be explicitly enabled in CI if this is wanted.