Fix lint shard to stop using `--lint-per-target-caching`
This breaks with MyPy because our BUILD files do not properly declare dependencies, so when running on a per-target basis, we have missing dependencies.

We could make `ci.py` much more complex to run MyPy separately from the other linters, but we instead take the simpler approach. A big benefit of removing this option is that the output will now be substantially more parse-able. We won't have hundreds of results anymore.
# Delete this line to force CI to run Clippy and the Rust tests.
[ci skip-rust-tests]
# Delete this line to force CI to run the JVM tests.
[ci skip-jvm-tests]