Automatically include `__init__.py` files (#10166)
### Problem
In Python, `__init__.py` files are significant. For example, before PEP 420, they are needed for namespace packages to work.

Right now, we only include the file if it is included in a target's `sources`. This is a problem because often ancestor files are significant. For example, we need the ancestor `__init__.py` files in `contrib/` for Pants to know that the entire package is a namespace package.

### Solution
Before generating, use `PathGlobs` to find any `__init__.py` files that are missing from the original snapshot. This will find as many as possible, and won't error on unmatched globs. The rest are then generated.

Closes https://github.com/pantsbuild/pants/issues/8082.

[ci skip-rust-tests]
[ci skip-jvm-tests]