Attempt to adjust file handle limits and warn if they cannot be adjusted (#10174)
### Problem

A few consecutive versions of macOS had particularly low "soft" `ulimit -n` (`NOFILE`: ie number of file handle) values.

### Solution

The soft/current limit can be raised as high the hard/max limit, and there is likely to be a gap between them. Attempt to raise the current limit to the maximum, and warn if we can't raise above a known-reasonable value.

### Result

In at least some cases, we will successfully, silently adjust the limit. In other cases we will warn the user that they have a value that is likely to cause trouble.

[ci skip-jvm-tests]