Fix `dependencies --transitive` to include target roots that are also dependencies (#10186)
Before, `./pants dependencies --transitive ::` would return empty input. This is surprising because many of the target roots are in fact dependencies of other target roots.

We change `TransitiveTargets` to have the field `dependencies`, which is any of the transitive dependencies of the target roots. The dependencies may include values from the target roots.

`TransitiveTargets` keeps a `closure` property to combine the roots with the dependencies.

[ci skip-rust-tests]