Remove --zipkin-trace-v2 option (#10184)
This commit removes the --zipkin-trace-v2 option and tests using it. In the v2 world, we will implement zipkin tracing (if necessary) with a plugin that consumes streaming workunits. This also allows us to remove the `get_workunits` rust FFI function, which was only used to implement this functionality, and no longer have to pass around the zipkin tracing boolean in rust code.