Remove v1 portions of `contrib/` infrastructure (#10177)
We no longer have any contrib plugins. For now, we are building in all functionality directly to `pantsbuild.pants` and having users activate via `backend_packages`.

Possibly, we will want to use a mechanism like contrib again in the future. If we decide to do so, we can always re-add this mechanism. In the meantime, we don't want to keep this cruft around.

[ci skip-rust-tests]