Refactor remoting options into separate struct (#10215)
### Problem

The large number of remoting-related options passed into the engine's `Core::new` function is unwieldy. Adding a new option is a chore.

### Solution

Introduce `RemotingOptions` struct and a `PyRemotingOptions` wrapper. Use both.

### Result

The remoting-related options are now grouped into a single RemotingOptions that can be passed around. It is clear where to add new remoting-related options.