Delete the old BinaryTool mechanism.
It's replaced by ExternalTool. This change switches
the last remaining users of BinaryTool over to ExternalTool,
and deletes the former.

Note that the native tool classes had extra utility
methods on them that assumed that tools were downloaded
locally. This is not the v2 way of doing things.
These utilities were deleted, and the native tool using code
will need to be rewritten to accept a digest+relpath,
in the v2 idiom.

[ci skip-rust-tests]