Fix `--changed-dependees` to work when v1 is disabled (#10235)
While we were using a rule to calculate the dependees, it was using a v1 code path based on TargetAdaptor. This means that `--changed-dependees` caused a crash when v1 is disabled.

To land this, we refactor `dependees` into two helper rules so that any rule can now `await Get(Dependees, DependeesRequest)`. 

[ci skip-rust-tests]