Fix remaining internal dependency cycles (#10254)
This requires a few fixes:
* Add an override `modules` parameter to `pants_requirement`. We use `pants_requirement` in one testproject, and this was causing dep inference to treat every Pants import as third party.
* Convert `bin/` to file-level targets to break a cycle with `pantsd/` and `bin/`.
* Remove type hint from `deprecated.py` to break a cycle betwen `base:deprecated` and `option/`.
* Remove type hint from `interactive_process.py` to break a cycle between `engine:interactive_process` and `engine/internals:scheduler`.
* Remove the AWS Lambda example. It causes there to be two owners of `pex`, which means that we no longer can infer it. We already have an example in the example repo.

We also turn off inference from inferred imports. This has too many false positives, and we should not use it until we have a) exclude syntax for dependencies and/or a global option to disable for certain strings, like what Toolchain has.

To avoid introducing new cycles, we enable dep inference. In a followup, we will remove `dependencies` from BUILD files.

[ci skip-rust-tests]