Lower the default max-memory usage of pantsd. (#10287)
### Problem

The default max memory usage from #10003 was chosen with larger monorepos in mind, and didn't match the expectation of consumers in smaller repos. 

### Solution

Very large repos will have folks who are able/willing to adjust limits like this to optimize for their repo size, so adjust the default down in favor of having good defaults.  This relates to #7675, but it isn't the time to dive in there.

### Result

Fixes #10264.

[ci skip-rust-tests]