remoting: default is streaming client (#10260)
### Problem

The streaming RE client is not the default. It has been enabled in CI for a few weeks now. New development is happening in the streaming client and not in the polling client. Time to make the streaming RE client the default.

### Solution

Set the streaming RE client as the default.

### Result

Nothing should break.