Unify `LegacyPythonCallbackParser` and `Parser` (#10281)
The original engine experiment allowed for multiple different parser techniques, such as a JSON parser and an assignment-based parser. As a result, the code for `parser.py` and related abstractions like `AddressMap` was kept very generic and abstract.

A few years later, we don't have any plans to move away from the current BUILD file callback-based syntax (rather, we're moving towards reducing the need for BUILD files in the first place). So, this code is now overly abstracted and slows us down.

While this PR does not fully remove the ability for different parsers, it makes several refactors to reduce the complexity of the Parser code path.

[ci skip-rust-tests]