Remove v1 `source/` abstractions like `EagerFileSet` (#10292)
We also rewrite what a `Filespec` typed dictionary looks like so that it's a flat list of includes and excludes globs, whereas the excludes could previously be a nested list. Why keep `Filespec`? We use this for `./pants filedeps --globs` and also for the `owner_of` mechanism to find owners without hydrating sources.

[ci skip-rust-tests]