Add option for specifying environment variable with concurrency slot (#10297)
### Problem

In some scenarios, pants users might want their tests to be aware of the amount of concurrency those tests are being run with. For instance, database-heavy Django tests might want to be able to reuse a limited number of databases, instead of spawning and destroying a new test database instance for every single concurrent test, which is resource-intensive.

### Solution

This PR introduces the ability to optionally create a variable in the environment of a `Process` that has a numerical ID corresponding to the "slot" that this `Process` is run under by the `BoundedCommandRunner` system. The name of this environment variable is controlled by a new pants option `--process-execution-slot-environment-variable`, and will not be present if this option is the empty string. 

The ID presented to processes via this variable comes from the `BoundedCommandRunner`, and will always be a small numerical value greater than 0 and less than or equal to the maximum amount of concurrency allowed by that `BoundedCommandRunner` instance (which in turn is set by the `--process-execution-{local,remote}-parallelism` options).