Only generate if original target has > 1 owning file
There is no need to generate if the target already has file-level precision. It's a little confusing to do so because it's not clear why Pants is doing this and which you should use.

Now, it will be much clearer when Pants is doing its magic to generate targets vs. when it's depending on traditional targets. Traditional dependencies will be anything explicitly specified in BUILD files _and_ anything inferred that is already file level. By removing that noise, you can much more easily see in `./pants dependencies` what is generated vs. not.

# Rust tests will be skipped. Delete if not intended.
[ci skip-rust-tests]