Refactor: add `generate_subtarget_address`
This both makes code more explicit and avoids unnecessary work. It seems now clear that the only place that will handle the actual creation of the target is the rule that goes from Address->Target. Everything else only needs to generate the address, and that rule will do its magic. So, it was not necessary to do the extra work of generating the target multiple times when we just needed the address.

Finally, by changing the parameter for `generate_subtarget_address` from `Target` to `Address`, this will make it much easier to integrate with file arguments. File arguments intentionally only use Address, so we previously would have had needed to hydrate targets unnecessarily.

# Rust tests will be skipped. Delete if not intended.
[ci skip-rust-tests]