If >1 owner of a file, do not use generated subtargets
This solves an ambiguous representation problem where we would show the address `demo.py` more than once, making it look like there's duplication when really they're different targets.

We must be careful that if one generated subtarget in the greater target is ambiguous, all the sibling generated targets should not be used.