Fix issue with SourcesSnapshot of the same files being duplicated by address specs and filesystem specs
Like with the AddressesWithOrigins rule, it is confusing for the same file to show up in two different contexts. Instead, we should detect when this happens and merge. With addresses, this merge means preferring the original explicit target over generated subtargets. With SourcesSnapshot, this simply means using MergeDigests.

As a result, cloc.py and source_file_validator.py are both simpler than before.

# Rust tests will be skipped. Delete if not intended.
[ci skip-rust-tests]