Allow specifying address and file arguments at the same time (#10345)
This is now possible:

```
▶ ./pants list src/python/pants/util:tests src/python/pants/util/strutil.py
src/python/pants/util:tests
src/python/pants/util:util
```

### Motivation

It was bad to inject two different Params conditionally into the root of the graph because this allows for divergent behavior. Closes https://github.com/pantsbuild/pants/issues/10152.

Further, this was blocking a change where file arguments will now use generated subtargets when possible. There, we need to convert `--changed-since` from the resolved addresses back to the Specs. Some of those need to be Filesystem Specs, if we used generated subtargets. Others need to be Address Specs. So we must be able to combine the two types of specs.

[ci skip-rust-tests]