Fix bug where workunit completion was not reported correctly (#10277)
### Problem

https://github.com/pantsbuild/pants/pull/10179 had the effect of changing the architecture of the workunit store so that when workunits were started and completed, the methods responsible for these tasks would push them onto a mpsc queue, and they would only actually be added to the store when one of the methods that handed them pulled the workunits off the queue. This is fine when the dynamic UI is running, because the `heavy_hitters` method will be running constantly, and therefore logging workunit completion messages. However, in the `--no-dynamic-ui` case, `heavy_hitters` doesn't run, and therefore nothing handles the workunits on the queue, 

### Solution

The workunit store is already aware of whether the dynamic UI is running or not, so this commit has the workunit completion method check this flag. If it's set, we send workunits over the mpsc queue, and expect them to be added to the store (a locking operation) by `heavy_hitters`, which needs to grab the lock anyway. When the dynamic UI is not running, the workunit add method doesn't use the heavy hitters mpsc queue at all, and just emits the display message for the workunit completion. Additionally, a new integration test is added to test that these workunit completion log messages get printed.


### Result

Fixes https://github.com/pantsbuild/pants/issues/10274