Ignore dead symlinks on Linux as well as OSX (#10339)
### Problem

The intended API of `Snapshot` is that dead symlinks are both watched and ignored, but currently on Linux we fail instead (in a fashion that does not give a useful error message). This turns out to be due to our use of the `notify` crate to watch a path before executing a syscall: in this case, `read_link`.

### Solution

Use a very small patch to the `notify` crate (see https://github.com/notify-rs/notify/issues/255) to avoid following symlinks for watches. Additionally, improve `watch` error messages to make this easier to track down in the future. Finally, add tests to cover the behavior.

### Result

Dead symlinks will be ignored (but also still watched, in case they "come back to life") uniformly across Linux and OSX.