Tweak `any_matches_filespec` to return which paths matched (#10358)
We need this extra detail to be able to know when calculating the owners of files whether to use the original owning target or a generated subtarget (https://github.com/pantsbuild/pants/pull/10356). This will allow us to determine if, for a particular target, we are matching deleted files vs. concrete files.

[ci skip-rust-tests]