Use colors for log messages (#10278)
### Problem

We would like the logging output of pants to be friendlier by using color judiciously.

### Solution

This commit adds text coloring and styling to the output of log messages if the `--colors` global option is set. To avoid overwhelming the screen with color, we only apply color to the bracketed log level signifier (e.g. `[INFO]`, `[DEBUG]`), except in the case of ERROR-level log messages, where we display all the text in red. The color scheme used is:

TRACE - magenta
DEBUG - green
INFO - default color,
WARN - red
ERROR - red