Allow dependencies on files in BUILD files (#10381)
Like with dep inference inferring file-level dependencies via generated subtargets, users can now do this explicitly in BUILD files. Users simply put the file name, and Pants finds the owning target(s). For relative files, you start with `./`, e.g. `./sibling.py`.

If there is not exactly 1 owner for the file, then we will error, whereas dep inference no-ops. This is to follow the principle of least surprise; we require users to disambiguate.

While we do not expect this to be used frequently, this is an important unification. Now, anything outputted by `./pants dependencies` can be explicitly put in a BUILD file and will work.

[ci skip-rust-tests]