remoting: use debug level for RE client workunits saved from remote (#10382)
### Problem

A recent bug fix causes Pants to print out INFO level messages even if the parent is DEBUG level. This now causes the workunits generated from remote execution timings to print out and spam the console with lots of messages like these: 

```
18:41:46.01 [INFO] Completed: remote execution worker output uploading
18:41:46.22 [INFO] Completed: remote execution action scheduling
18:41:46.22 [INFO] Completed: remote execution worker input fetching
18:41:46.22 [INFO] Completed: remote execution worker command executing
18:41:46.22 [INFO] Completed: remote execution worker output uploading
18:41:46.32 [INFO] Completed: remote execution action scheduling
18:41:46.32 [INFO] Completed: remote execution worker input fetching
```

### Solution

Set the workunits to DEBUG level.

### Result

Log should not be spammed.