Error if an exclude is not used
Also, generify the code so that we stop assuming dep inference uses generated subtargets and stop assuming that dependency injection uses normal Addresess. This is how it's implemented now, but it does not need to be. For example, today we mentioned how Java may not always want to use generated subtargets. Either way, this results in simpler code with fewer special cases.

[ci skip-rust-tests]