Allow ignoring dependencies with `!` (#10385)
It would be silly to do `[':sibling', '!:sibling]`, but this is useful for when dependency inference results in a false positive.

The user may either specify a normal address or they may specify a file, which will cause us to look for the owning target and to generate a subtarget. We validate that files have exactly one owner, like what we do with explicit file deps in general.

Every ignore must actually cause a dependency to be ignored; we do not allow unmatched ignores. This reduces the risk of users thinking something is being ignored when it's really not.

[ci skip-rust-tests]