Fix validation of dependencies ignores breaking with generated subtargets (#10407)
### Problem
We eagerly validate that every `!` is used in the `dependencies` field as a matter of good UX, because it's confusing how all the different pieces fit together.

However, imagine this: you have a target that owns 5 files. You set on that target `!./bad.py`. However, only one of those 5 files actually uses `bad.py`. This means that the 4 other generated subtargets would now all complain that the ignore `!./bad.py` is unused.

[ci skip-rust-tests]