Print initial logs from pantsd using local time format
### Problem

In the Rust logging framework, we set up a custom formatting string for logs being sent to the dynamic UI's log-printing framework, and configured this custom code to print a timestamp in the local time zone. However, pantsd logs don't use the dynamic UI's log-printing framework, but rather prints logs via the `MaybeWriteLogger` mechanism which wraps `WriteLogger`.

### Solution

Ensure that the `WriteLogger` configuration that pantsd does use has localtime set. 

This does not make the dynamic-UI logs completely identical to the `WriteLogger` logs, because the dynamic-UI code has a bit of custom logic to print 2 decimal points of second precision in the timestamp, but the time format string logic that `WriteLogger` configuration uses does not allow this. Unfortunately, short of making upstream changes to the time formatting library, there doesn't seem to be an easy way to change this.

### Result

(mostly) fixes https://github.com/pantsbuild/pants/issues/10128