Add a hint about adjusting max_user_watches on Linux (#10413)
### Problem

ENOSPC errors on Linux in the context of file watching tend to be caused by bumping into the maximum number of file watches. See #10408.

### Solution

Enrich the ENOSPC error with a link to the docs and attempt to report the current limit by reading proc.

### Result

Testing in an Ubuntu VM with a low watch count reports the enriched error. Fixes #10408.