Port the plugin resolver test to v2 test style.
Previously it used some hackery to get hold of various
subsystems to construct a pex to run setup.py, in order
to create the plugin dists it then tested against.

This change uses v2 test idioms to create and run
that pex.

As a result, we can delete a lot of code that only
hung around because of this one test.

Unfortunately, this required making the test functions
methods of a test class, meaning we can no longer
use pytest.mark.parameterize, and have to "parameterize"
manually.

[ci skip-rust-tests]