Rename argument in streaming workunit callback (#10432)
### Problem

The streaming workunit system invokes the callback with several named keyword arguments, one of which is `workunits`. We later added a `started_workunits` kwarg intended for started workunits, with `workunits` now containing only finished workunits. This is confusing since it implies that the `workunits` argument contains all the workunits rather than only the ones that have completed.

### Solution

Rename the `workunits` kwarg to `completed_workunits`. This breaks any streaming workunit clients not updated to expect a `completed_workunit` argument or that *do* expect a `workunit` argument, but that should be okay in the context of the upcoming 2.0 major release bump.