Rename `Workspace.materialize_directory()` to `Workspace.write_digest()` (#10448)
We've had several people mention that `materialize_directory` confused them because a) `directory` suggests you can only materialize a directory and not a file, and b) `materialize` isn't very simple. Instead, `write_digest()` fits with the general naming conventions we've established for the file system: https://www.pantsbuild.org/v2.0/docs/rules-api-file-system

This also removes the parallel `Workspace.materialize_directories()`. It required that each distinct digest have zero overlapping contents, which is not very common. We never used this method in production. Instead, users should call `await Get(Digest, MergeDigests)` beforehand.

`write_digest()` no longer returns any result, as the value was redundant with the contents of the input `digest`. If the user needs the names of the files written, they can lift the input Digest into a Snapshot and use `snapshot.files`.

Finally, we remove the intermediate type `WriteDigest`, as there is no need for this and it suggests that you may use the type in rules, when you cannot.