Unify accessing options for a `GoalSubsystem` with `Subsystem` (#10456)
With a normal `Subsystem`, you use `self.options.foo` to access an option value. But with a `GoalSubsystem`, you use `self.values.foo`. This difference is confusing and requires unnecessary context switching.

Instead, we rename `.values` to `.options`. 

We also stop using the naming convention of calling the class `MyGoalOptions` and instead use `MyGoalSubsystem`. This makes things more consistent.

In a later followup, it would be good to remove the intermediate types of `OptionableFactory` and `SubsystemClientMixin` to have `GoalSubsystem` instead directly subclass `Subsystem`.

[ci skip-rust-tests]